Fix claim tax payment bug
When claiming a pixel, the claimer can receive some funds in addition to
the pixel. If the account is indebted, we have to pay taxes from this,
similarly to what we do with a normal deposit.

To fix this, we added the tax payment calculation from the deposit
function into the increase balance helper. So we now handle taxes
properly whenever the balance is increased.

Besides during depositing and claiming, the balance is also increased
during selling. There however wasn't a bug and the logic isn't changed
because either
- the account has a positive balance and no taxes have to be paid
- the account has a negative balance and the buy price is zero, so no
  taxes can be paid